backspace()移动至insert.rs
